Move area_difference to test utils
