+++ tests/test_equilibrium.py
@@ -3,9 +3,9 @@ import numpy as np
 from netCDF4 import Dataset
 import pytest
 
+from .utils import area_difference, compute_coords
 from desc.equilibrium import EquilibriaFamily, Equilibrium
 from desc.grid import Grid, LinearGrid
-from desc.utils import area_difference
 from desc.__main__ import main
 from desc.geometry import ZernikeRZToroidalSection
 from desc.objectives import (
@@ -90,68 +90,13 @@ def test_compute_flux_coords(SOLOVEV):
     np.testing.assert_allclose(nodes, flux_coords, rtol=1e-5, atol=1e-5)
 
 
-def _compute_coords(equil, check_all_zeta=False):
-
-    if equil.N == 0 and not check_all_zeta:
-        Nz = 1
-    else:
-        Nz = 6
-
-    Nr = 10
-    Nt = 8
-    num_theta = 1000
-    num_rho = 1000
-
-    # flux surfaces to plot
-    rr = np.linspace(0, 1, Nr)
-    rt = np.linspace(0, 2 * np.pi, num_theta)
-    rz = np.linspace(0, 2 * np.pi / equil.NFP, Nz, endpoint=False)
-    r_grid = LinearGrid(rho=rr, theta=rt, zeta=rz)
-
-    # straight field-line angles to plot
-    tr = np.linspace(0, 1, num_rho)
-    tt = np.linspace(0, 2 * np.pi, Nt, endpoint=False)
-    tz = np.linspace(0, 2 * np.pi / equil.NFP, Nz, endpoint=False)
-    t_grid = LinearGrid(rho=tr, theta=tt, zeta=tz)
-
-    # Note: theta* (also known as vartheta) is the poloidal straight field-line
-    # angle in PEST-like flux coordinates
-
-    # find theta angles corresponding to desired theta* angles
-    v_grid = Grid(equil.compute_theta_coords(t_grid.nodes))
-    r_coords = equil.compute("R", r_grid)
-    v_coords = equil.compute("Z", v_grid)
-
-    # rho contours
-    Rr1 = r_coords["R"].reshape(
-        (r_grid.num_theta, r_grid.num_rho, r_grid.num_zeta), order="F"
-    )
-    Rr1 = np.swapaxes(Rr1, 0, 1)
-    Zr1 = r_coords["Z"].reshape(
-        (r_grid.num_theta, r_grid.num_rho, r_grid.num_zeta), order="F"
-    )
-    Zr1 = np.swapaxes(Zr1, 0, 1)
-
-    # vartheta contours
-    Rv1 = v_coords["R"].reshape(
-        (t_grid.num_theta, t_grid.num_rho, t_grid.num_zeta), order="F"
-    )
-    Rv1 = np.swapaxes(Rv1, 0, 1)
-    Zv1 = v_coords["Z"].reshape(
-        (t_grid.num_theta, t_grid.num_rho, t_grid.num_zeta), order="F"
-    )
-    Zv1 = np.swapaxes(Zv1, 0, 1)
-
-    return Rr1, Zr1, Rv1, Zv1
-
-
 @pytest.mark.slow
 def test_to_sfl(SOLOVEV):
 
     eq = EquilibriaFamily.load(load_from=str(SOLOVEV["desc_h5_path"]))[-1]
 
-    Rr1, Zr1, Rv1, Zv1 = _compute_coords(eq)
-    Rr2, Zr2, Rv2, Zv2 = _compute_coords(eq.to_sfl())
+    Rr1, Zr1, Rv1, Zv1 = compute_coords(eq)
+    Rr2, Zr2, Rv2, Zv2 = compute_coords(eq.to_sfl())
     rho_err, theta_err = area_difference(Rr1, Rr2, Zr1, Zr2, Rv1, Rv2, Zv1, Zv2)
 
     np.testing.assert_allclose(rho_err, 0, atol=2.5e-5)
