+++ tests/test_examples.py
@@ -13,8 +13,8 @@ from desc.objectives import (
     FixIota,
     FixPsi,
 )
-from desc.vmec import VMECIO
 from desc.vmec_utils import vmec_boundary_subspace
+from .utils import area_difference_vmec
 
 
 def test_SOLOVEV_vacuum(SOLOVEV_vac):
@@ -31,7 +31,7 @@ def test_SOLOVEV_results(SOLOVEV):
     """Tests that the SOLOVEV example gives the same result as VMEC."""
 
     eq = EquilibriaFamily.load(load_from=str(SOLOVEV["desc_h5_path"]))[-1]
-    rho_err, theta_err = VMECIO.area_difference_vmec(eq, SOLOVEV["vmec_nc_path"])
+    rho_err, theta_err = area_difference_vmec(eq, SOLOVEV["vmec_nc_path"])
 
     np.testing.assert_allclose(rho_err, 0, atol=1e-3)
     np.testing.assert_allclose(theta_err, 0, atol=1e-5)
@@ -41,7 +41,7 @@ def test_DSHAPE_results(DSHAPE):
     """Tests that the DSHAPE example gives the same result as VMEC."""
 
     eq = EquilibriaFamily.load(load_from=str(DSHAPE["desc_h5_path"]))[-1]
-    rho_err, theta_err = VMECIO.area_difference_vmec(eq, DSHAPE["vmec_nc_path"])
+    rho_err, theta_err = area_difference_vmec(eq, DSHAPE["vmec_nc_path"])
 
     np.testing.assert_allclose(rho_err, 0, atol=2e-3)
     np.testing.assert_allclose(theta_err, 0, atol=1e-5)
@@ -51,7 +51,7 @@ def test_HELIOTRON_results(HELIOTRON):
     """Tests that the HELIOTRON example gives the same result as VMEC."""
 
     eq = EquilibriaFamily.load(load_from=str(HELIOTRON["desc_h5_path"]))[-1]
-    rho_err, theta_err = VMECIO.area_difference_vmec(eq, HELIOTRON["vmec_nc_path"])
+    rho_err, theta_err = area_difference_vmec(eq, HELIOTRON["vmec_nc_path"])
 
     np.testing.assert_allclose(rho_err.mean(), 0, atol=1e-2)
     np.testing.assert_allclose(theta_err.mean(), 0, atol=2e-2)
