+++ tests/test_profiles.py
@@ -4,7 +4,7 @@ import pytest
 from desc.io import InputReader
 from desc.profiles import PowerSeriesProfile
 from desc.equilibrium import Equilibrium
-from tests.test_equilibrium import _compute_coords, area_difference
+from .utils import compute_coords, area_difference
 
 
 class TestProfiles(unittest.TestCase):
@@ -22,8 +22,8 @@ class TestProfiles(unittest.TestCase):
         eq1.solve()
         eq2.solve()
 
-        Rr1, Zr1, Rv1, Zv1 = _compute_coords(eq1, check_all_zeta=True)
-        Rr2, Zr2, Rv2, Zv2 = _compute_coords(eq2, check_all_zeta=True)
+        Rr1, Zr1, Rv1, Zv1 = compute_coords(eq1, check_all_zeta=True)
+        Rr2, Zr2, Rv2, Zv2 = compute_coords(eq2, check_all_zeta=True)
         rho_err, theta_err = area_difference(Rr1, Rr2, Zr1, Zr2, Rv1, Rv2, Zv1, Zv2)
         np.testing.assert_allclose(rho_err, 0, atol=1e-7)
         np.testing.assert_allclose(theta_err, 0, atol=2e-11)
